Fix issue where mini view didn't update when pasting model data. Create map mini view. Rework map service to work properly with the workspace system.

+++ RanseiLink.View3D/SceneRenderer.cs
@@ -129,7 +129,7 @@ public class SceneRenderer : ISceneRenderer
         var nsbmd = LoadMaterialsFromPac(pac);
         models.Add(new ModelInfo(nsbmd));
 
-        var pslm = _mapService.Retrieve(mapId);
+        var pslm = _mapService.Retrieve((int)mapId);
         var mapGridHeight = pslm.TerrainSection.MapMatrix.Count;
         var mapGridWidth = pslm.TerrainSection.MapMatrix[0].Count;
